% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecsymr-package.R
\docType{package}
\name{vecsymr-package}
\alias{vecsymr}
\alias{vecsymr-package}
\title{vecsymr: Vector Symbolic Architecture (VSA) primitives for experimentation}
\description{
Implements Vector Symbolic Architecture (VSA) primitives to support experimentation. Design choices are Ross Gayler's personal preferences for a simple experimental system (the VSA equivalent of the geneticist's fruitfly). Vector elements are phasors (unit magnitude complex numbers). There are some extra features intended to support non-negativity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgayler/vecsymr}
  \item Report bugs at \url{https://github.com/rgayler/vecsymr/issues}
}

}
\author{
\strong{Maintainer}: Ross Gayler \email{ross@rossgayler.com} (\href{https://orcid.org/0000-0003-4679-585X}{ORCID})

}
\keyword{internal}
