% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsa_mk_atom_bipolar.R
\name{vsa_mk_atom_bipolar}
\alias{vsa_mk_atom_bipolar}
\title{Make an atomic VSA vector (bipolar)}
\usage{
vsa_mk_atom_bipolar(vsa_dim, seed = NULL)
}
\arguments{
\item{vsa_dim}{An integer (>0) - The dimensionality of the returned VSA vector.}

\item{seed}{An integer or NULL - The random generator seed for the returned VSA vector.}
}
\value{
A numeric vector with length vsa_dim.
}
\description{
Make an atomic VSA vector (bipolar)
}
\examples{
vsa_mk_atom_bipolar(10L)
}
